refactor: use colorama for colorful logging
