+++ mr_robot/__main__.py
@@ -1,88 +1,18 @@
 import asyncio
-import atexit
-import json
 import logging.config
 import logging.handlers
 import signal
 import sys
 from pathlib import Path
 
-import coloredlogs
 import disnake
 import httpx
-from dotenv import load_dotenv
 from sqlalchemy import event
 from sqlalchemy.engine import Engine
 
 from mr_robot.bot import MrRobot
 from mr_robot.constants import Client, Database
-
-load_dotenv()
-
-
-def setup_logging_modern() -> None:
-    with open(Client.logging_config_file, "r") as file:
-        config = json.load(file)
-    logging.config.dictConfig(config)
-    queue_handler = logging.getHandlerByName("queue_handler")
-    if queue_handler is not None:
-        queue_handler.listener.start()  # type: ignore[reportAttributeAccessIssue]
-        atexit.register(queue_handler.listener.stop)  # type: ignore[reportAttributeAccessIssue]
-
-
-def setup_logging() -> None:
-    root_logger = logging.getLogger()
-
-    log_file = Path(Client.log_file_name)
-    log_file.parent.mkdir(exist_ok=True)
-
-    formatter = "[ %(levelname)s | %(name)s | %(module)s | L%(lineno)d ] %(asctime)s: %(message)s"
-    file_formatter = logging.Formatter(
-        "[ %(levelname)s | %(name)s | %(module)s | %(funcName)s | %(filename)s | L%(lineno)d ] %(asctime)s: %(message)s"
-    )
-
-    file_handler = logging.handlers.RotatingFileHandler(
-        log_file,
-        mode="a",
-        # File handler rotates log every 5 MB
-        maxBytes=5 * (2**20),
-        backupCount=10,
-    )
-    file_handler.setFormatter(file_formatter)
-    root_logger.addHandler(file_handler)
-
-    # console_handler = logging.StreamHandler()
-    # console_handler.setFormatter(formatter)
-    # root_logger.addHandler(console_handler)
-
-    file_handler.setLevel(logging.DEBUG)
-    # console_handler.setLevel(logging.INFO)
-
-    coloredlogs.DEFAULT_LEVEL_STYLES = {
-        **coloredlogs.DEFAULT_LEVEL_STYLES,
-        "critical": {"color": "red"},
-        "warning": {"color": "yellow"},
-        "info": {"color": "green"},
-        "debug": {"color": "white"},
-    }
-    coloredlogs.install(
-        level=logging.DEBUG, stream=sys.stdout, logger=root_logger, fmt=formatter
-    )
-
-    root_logger.setLevel(logging.DEBUG)
-    logging.getLogger("httpx").setLevel(logging.WARNING)
-    logging.getLogger("mafic").setLevel(logging.WARNING)
-    logging.getLogger("httpcore").setLevel(logging.WARNING)
-    logging.getLogger("disnake").setLevel(logging.INFO)
-    logging.getLogger("sqlalchemy.engine").setLevel(logging.WARNING)
-    logging.getLogger("core").setLevel(logging.WARNING)
-    logging.getLogger("sqlalchemy.engine.Engine").setLevel(logging.WARNING)
-    logging.getLogger("sqlalchemy.pool").setLevel(logging.WARNING)
-    logging.getLogger("sqlalchemy.orm").setLevel(logging.WARNING)
-    logging.getLogger("aiosqlite").setLevel(logging.WARNING)
-    logging.getLogger("streamlink").disabled = True
-
-    root_logger.info("Logger Initialized!")
+from mr_robot.log import setup_logging
 
 
 @event.listens_for(Engine, "connect")
